Removed example project and replace with integrationTests
