--- knotx-example/knotx-example-action-adapter-http/src/test/resources/first-response.json
@@ -1,11 +1,0 @@
-{
-  "message": "this is webservice no. 1",
-  "welcomeMessage": "welcome from nr. 1",
-  "responseMessage": "message received",
-  "body": {
-    "a": "message - a",
-    "b": "message - b",
-    "c": "message - c"
-  },
-  "statistics": 20
-}
