--- knotx-example/knotx-example-action-adapter-http/src/test/resources/knotx-action-adapter-http-test.json
@@ -1,44 +1,0 @@
-{
-  "modules": [
-    "actionAdapter=java:io.knotx.adapter.action.http.HttpActionAdapterVerticle",
-    "serviceMock=java:io.knotx.mocks.MockServiceVerticle"
-  ],
-  "config": {
-    "actionAdapter": {
-      "options": {
-        "config": {
-          "address": "knotx.adapter.action.http",
-          "clientOptions": {
-            "maxPoolSize": 1000,
-            "keepAlive": false
-          },
-          "services": [
-            {
-              "path": "/service/mock/.*",
-              "domain": "localhost",
-              "port": 4000,
-              "allowedRequestHeaders": [
-                "Content-Type",
-                "X-*"
-              ]
-            },
-            {
-              "path": "/service/.*",
-              "domain": "localhost",
-              "port": 8080
-            }
-          ]
-        }
-      }
-    },
-    "serviceMock": {
-      "options": {
-        "config": {
-          "mockDataRoot": "mock/service",
-          "bouncing": true,
-          "httpPort": 4000
-        }
-      }
-    }
-  }
-}
