--- knotx-example/knotx-example-app/config/application.conf
@@ -1,120 +1,0 @@
-########### Modules to start ###########
-modules = [
-  "server=io.knotx.server.KnotxServerVerticle"
-  "httpRepo=io.knotx.repository.http.HttpRepositoryConnectorVerticle"
-  "fsRepo=io.knotx.repository.fs.FilesystemRepositoryConnectorVerticle"
-  "splitter=io.knotx.splitter.FragmentSplitterVerticle"
-  "assembler=io.knotx.assembler.FragmentAssemblerVerticle"
-  "hbsKnot=io.knotx.knot.templating.HandlebarsKnotVerticle"
-  "serviceKnot=io.knotx.knot.service.ServiceKnotVerticle"
-  "actionKnot=io.knotx.knot.action.ActionKnotVerticle"
-  "serviceAdapter=io.knotx.adapter.service.http.HttpServiceAdapterVerticle"
-  "actionAdapter=io.knotx.adapter.action.http.HttpActionAdapterVerticle"
-  "repoMock=io.knotx.mocks.MockRemoteRepositoryVerticle"
-  "serviceMock=io.knotx.mocks.MockServiceVerticle"
-  "gatewayKnot=io.knotx.gateway.GatewayKnotVerticle"
-  "reqProcessorKnot=io.knotx.gateway.RequestProcessorKnotVerticle"
-  "respProviderKnot=io.knotx.gateway.ResponseProviderKnotVerticle"
-]
-
-########### Globals ###########
-global {
-  serverPort = 8092
-
-  snippetTagName = script
-
-  address {
-    httpRepo = knotx.core.repository.http
-    fsRepo = knotx.core.repository.filesystem
-
-    knot {
-      service = knotx.knot.service
-      hbs = knotx.knot.handlebars
-      action = knotx.knot.action
-      gateway = knotx.gateway.gatewayknot
-      exampleReqProcessor = knotx.gateway.requestprocessor
-    }
-
-    adapter {
-      action = knotx.adapter.action.http
-      service = knotx.adapter.service.http
-    }
-  }
-
-  httpRepo {
-    domain = localhost
-    port = 3001
-  }
-
-  httpService {
-    domain = localhost
-    port = 3000
-  }
-}
-
-########### Modules configurations ###########
-config.server {
-  options.config {
-    include required("includes/server.conf")
-  }
-}
-
-config.httpRepo {
-  options.config {
-    include required("includes/httpRepo.conf")
-  }
-}
-
-config.fsRepo {
-  options.config {
-    include required("includes/fsRepo.conf")
-  }
-}
-
-config.splitter.options.config {
-  snippetTagName = ${global.snippetTagName} # Do not change
-}
-
-config.assembler.options.config {
-  snippetTagName = ${global.snippetTagName} # Do not change
-}
-
-config.hbsKnot {
-  options.config.address = ${global.address.knot.hbs}
-  options.config.cacheSize = 1000
-}
-
-config.serviceKnot {
-  options.config {
-    include required("includes/serviceKnot.conf")
-  }
-}
-
-config.serviceAdapter {
-  options.config {
-    include required("includes/serviceAdapter.conf")
-  }
-}
-
-config.actionKnot {
-  options.config {
-    include required("includes/actionKnot.conf")
-  }
-}
-
-config.actionAdapter {
-  options.config {
-    include required("includes/actionAdapter.conf")
-  }
-}
-
-config.repoMock.options.config {
-  mockDataRoot = mock/repository
-  httpPort = ${global.httpRepo.port}
-}
-
-config.serviceMock.options.config {
-  mockDataRoot = mock/service
-  bouncing = true
-  httpPort = ${global.httpService.port}
-}
