--- knotx-example/knotx-example-app/config/bootstrap.json
@@ -1,14 +1,0 @@
-{
-  "configRetrieverOptions": {
-    "scanPeriod": 5000,
-    "stores": [
-      {
-        "type": "file",
-        "format": "conf",
-        "config": {
-          "path": "${KNOTX_HOME}/config/application.conf"
-        }
-      }
-    ]
-  }
-}
