--- knotx-example/knotx-example-app/config/includes/actionAdapter.conf
@@ -1,33 +1,0 @@
-address = ${global.address.adapter.action}
-
-clientOptions {
-  maxPoolSize = 1000
-  keepAlive = false
-}
-
-customRequestHeader {
-  name = Server-User-Agent
-  value = Knot.x
-}
-
-services = [
-  {
-    path = "/service/mock/.*"
-    domain = ${global.httpService.domain}
-    port = ${global.httpService.port}
-    allowedRequestHeaders = [
-      "Content-Type"
-      "X-*"
-    ]
-  }
-  {
-    path = "/service/.*"
-    domain = localhost
-    port = 8080
-    allowedRequestHeaders = [
-      "Content-Type"
-      "X-*"
-    ]
-  }
-]
-
