--- knotx-example/knotx-example-app/config/includes/actionKnot.conf
@@ -1,67 +1,0 @@
-# Event bus address of the Action Knot
-address = ${global.address.knot.action}
-
-# Array of adapters that will be responsible for communicating with external services in order to
-# process client requests (form submisions).
-adapters = [
-  {
-    name = subscribe-competition
-    address = ${global.address.adapter.action}
-    params.path = /service/mock/post-competition.json
-    allowedRequestHeaders = [
-      Cookie
-    ]
-    allowedResponseHeaders = [
-      Set-Cookie
-      Location
-    ]
-  }
-  {
-    name = subscribe-newsletter
-    address = ${global.address.adapter.action}
-    params.path = /service/mock/post-newsletter.json
-    allowedRequestHeaders = [
-      Cookie
-    ]
-    allowedResponseHeaders = [
-      Set-Cookie
-      Location
-    ]
-  }
-  {
-    name = step1
-    address = ${global.address.adapter.action}
-    params.path = /service/mock/post-step-1.json
-    allowedRequestHeaders = [
-      Cookie
-      Content-Type
-      Content-Length
-    ]
-    allowedResponseHeaders = [
-      Set-Cookie
-      Location
-    ]
-  }
-  {
-    name = step2
-    address = ${global.address.adapter.action}
-    params.path = /service/mock/post-step-2.json
-    allowedRequestHeaders = [
-      Cookie
-      Content-Type
-      Content-Length
-    ]
-    allowedResponseHeaders = [
-      Set-Cookie
-      Location
-    ]
-  }
-]
-
-# Name of the hidden input tag which is added by Action Knot.
-formIdentifierName = _frmId
-
-# Vert.x event bus delivery options used when communicating with other verticles
-# see http://vertx.io/docs/vertx-core/dataobjects.html#DeliveryOptions for the details what can be configured
-#
-# deliveryOptions {}
