--- knotx-example/knotx-example-app/config/includes/fsRepo.conf
@@ -1,8 +1,0 @@
-# Event bus address on which the HTTP repository connector listens on. Default is 'knotx.core.repository.filesystem'
-#
-address = ${global.address.fsRepo}
-
-# Set the root folder of the repository on file system.
-# If catalogue equals empty string a verticle will look for the files in classpath
-#
-# catalogue = ""
