--- knotx-example/knotx-example-app/config/includes/serviceAdapter.conf
@@ -1,59 +1,0 @@
-# Event bus address of the Basic HTTP Service Adapter
-address = ${global.address.adapter.service}
-
-clientOptions {
-  maxPoolSize = 1000
-  idleTimeout = 120 # seconds
-}
-
-services = [
-  {
-    # A regexp.
-    # Any request to the service made by the adapter, is being made to the service with a given
-    # physical conenction details, only if the given path matching this path regexp
-    #
-    path = "/.*"
-
-    # A domain or IP of actual HTTP service the adapter will talk to
-    #
-    domain = ${global.httpService.domain}
-
-    # HTTP port of the service
-    #
-    port = ${global.httpService.port}
-
-    # List of request headers that will be send to the given service endpoint.
-    # Each header can be use wildcards '*' to generalize list
-    # E.g.
-    # allowedRequestHeaders = [
-    #     "X-*"
-    # ]
-    #
-    allowedRequestHeaders = [
-      ".*"
-    ]
-
-    # Additional request query parameters to be send in each request.
-    # E.g.
-    # queryParams {
-    #    type = books
-    #    query = java
-    # }
-    #
-    queryParams {}
-
-    # Additional headers to be send in each request to the service
-    # E.g:
-    # additionalHeaders {
-    #     someToken = abxcdsafsfdfes
-    # }
-    #
-    additionalHeaders {}
-  }
-]
-
-# Statically defined HTTP request header sent in every request to any service
-customHttpHeader {
-  name = X-User-Agent
-  value = Knot.x
-}
