--- knotx-example/knotx-example-app/config/includes/serviceKnot.conf
@@ -1,31 +1,0 @@
-# Event bus address of the handlebars Knot
-address = ${global.address.knot.service}
-
-# An
-services = [
-  {
-    name = first-service
-    address = ${global.address.adapter.service}
-    params.path = /service/mock/first.json
-    cacheKey = first
-  }
-  {
-    name = second-service
-    address = ${global.address.adapter.service}
-    params.path = /service/mock/second.json
-  }
-  {
-    name = third-service
-    address = ${global.address.adapter.service}
-    params.path = /service/mock/third.json
-  }
-  {
-    name = labelsRepository
-    address = ${global.address.adapter.service}
-  }
-]
-
-# Vert.x event bus delivery options used when communicating with other verticles
-# see http://vertx.io/docs/vertx-core/dataobjects.html#DeliveryOptions for the details what can be configured
-#
-# deliveryOptions {}
