--- knotx-example/knotx-example-app/src/test/resources/multipleFormWithPostResult2.html
@@ -1,93 +1,0 @@
-<!--
-
-    Copyright (C) 2016 Cognifide Limited
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-        http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-
--->
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="utf-8">
-  <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <link rel="shortcut icon" href="http://cognifide.com/favicon.ico?v=2"/>
-  <title>Knot.x</title>
-  <link href="https://bootswatch.com/simplex/bootstrap.min.css" rel="stylesheet"/>
-</head>
-<body>
-<div class="container-fluid">
-  <div class="row">
-    <div class="col-md-12">
-      <div class="jumbotron">
-        <h2>
-          Hello Knot.x - reactive micro-service assembler world!
-        </h2>
-        <p>
-          This template is served from the <strong>local</strong> repository.
-        </p>
-        <p>
-          <a class="btn btn-primary btn-large"
-             href="https://github.com/Cognifide/knotx">Learn more</a>
-        </p>
-      </div>
-    </div>
-  </div>
-  <div class="row">
-    <div class="col-md-12">
-       <h1>welcome in competition</h1>  <p>Please subscribe to our new competition:</p>
-      <form method="post" class="form-inline">
-        <input type="hidden" name="_frmId" value="competition" />
-        <div class="form-group">
-          <label for="name">Name</label>
-          <input type="text" name="name" id="name" />
-        </div>
-        <div class="form-group">
-          <label for="email">Email</label>
-          <input type="email" name="email" id="email" />
-        </div>
-        <button type="submit" class="btn btn-default">Submit</button>
-      </form>
-    </div>
-  </div>
-  <div class="row">
-    <div class="col-md-12">
-      <div id="newsletter-form-wrapper">
-        <h1>subscribe to newsletter</h1>
-        <p>thank you for subscribing to newsletter</p>
-        <p>Submitted: email@example.com</p>
-      </div>
-    </div>
-  </div>
-</div>
-<script src="https://code.jquery.com/jquery-2.2.4.min.js"
-        integrity="sha256-BbhdlvQf/xTY9gja0Dq3HiwQF8LaCRTXxZKRutelT44="
-        crossorigin="anonymous"></script>
-<script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"
-        integrity="sha384-Tc5IQib027qvyjSMfHjOMaLkfuWVxZxUPnCJA7l2mCWNIpG9mGCD8wGNIcPD7Txa"
-        crossorigin="anonymous"></script>
-<script type="text/javascript">
-    $(document).ready(function () {
-        function processForm(e) {
-            $.post("", $('#newsletter').serialize(), function (data, textStatus, jqXHR) {
-                $("#newsletter-form-wrapper").html(data);
-            });
-            e.preventDefault();
-        }
-
-        $('#newsletter').submit(processForm);
-    });
-
-</script>
-</body>
-</html>
