+++ knotx-it-tests/src/test/resources/logback-test.xml
@@ -1,0 +1,35 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+
+    Copyright (C) 2016 Cognifide Limited
+
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+        http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+
+-->
+<configuration debug="true">
+  <statusListener class="ch.qos.logback.core.status.OnConsoleStatusListener"/>
+
+  <property name="CONSOLE_KNOTX_LOG_PATTERN" value="%d{${KNOTX_LOG_DATEFORMAT_PATTERN:-yyyy-MM-dd HH:mm:ss.SSS}} [%thread] %-5level %logger{36} - %msg%n"/>
+  <include resource="io/knotx/logging/logback/console-appender.xml" />
+
+  <root level="INFO" additivity="false">
+    <appender-ref ref="CONSOLE"/>
+  </root>
+
+  <logger name="io.netty" level="ERROR" additivity="false"/>
+  <logger name="io.vertx" level="ERROR" additivity="false"/>
+
+  <logger name="io.knotx" level="TRACE" additivity="false">
+    <appender-ref ref="CONSOLE"/>
+  </logger>
+</configuration>
