+++ knotx-it-tests/src/test/resources/multipleFormWithAjaxPostResult.html
