+++ knotx-it-tests/src/test/resources/remoteSimpleResult.html
